+++ webapp/javascript/components/CustomDatePicker.tsx
@@ -5,7 +5,7 @@ import Button from '@webapp/ui/Button';
 import { formatAsOBject, getUTCdate } from '@webapp/util/formatDate';
 import useTimeZone from '@webapp/hooks/timeZone.hook';
 import Select from '@webapp/ui/Select';
-
+import TextField from '@webapp/ui/Form/TextField';
 import styles from './CustomDatePicker.module.scss';
 
 interface CustomDatePickerProps {
@@ -71,7 +71,6 @@ function CustomDatePicker({ from, until, onSubmit }: CustomDatePickerProps) {
     <div className="drp-custom">
       <h4>Custom Date Range</h4>
       <div className="from">
-        <label htmlFor="datepicker-from">From: </label>
         <DatePicker
           id="datepicker-from"
           selected={selectFromAsDate}
@@ -80,11 +79,16 @@ function CustomDatePicker({ from, until, onSubmit }: CustomDatePickerProps) {
           showTimeSelect
           startDate={selectFromAsDate}
           dateFormat="yyyy-MM-dd hh:mm aa"
-          className={styles.datepicker}
+          customInput={
+            <TextField
+              className={styles.datePickerInput}
+              label="From:"
+              variant="light"
+            />
+          }
         />
       </div>
       <div className="until">
-        <label htmlFor="datepicker-until">Until: </label>
         <DatePicker
           id="datepicker-until"
           selected={selectUntilAsDate}
@@ -95,12 +99,18 @@ function CustomDatePicker({ from, until, onSubmit }: CustomDatePickerProps) {
           endDate={selectUntilAsDate}
           minDate={selectFromAsDate}
           dateFormat="yyyy-MM-dd hh:mm aa"
-          className={styles.datepicker}
+          customInput={
+            <TextField
+              className={styles.datePickerInput}
+              label="Until:"
+              variant="light"
+            />
+          }
         />
       </div>
       {warning && <p style={{ color: 'red' }}>Warning: invalid date Range</p>}
 
-      <Button type="submit" kind="primary" onClick={() => updateDateRange()}>
+      <Button type="submit" kind="secondary" onClick={() => updateDateRange()}>
         Apply range
       </Button>
 
@@ -112,6 +122,7 @@ function CustomDatePicker({ from, until, onSubmit }: CustomDatePickerProps) {
           id="select-timezone"
           value={String(offset)}
           disabled={timeZoneOptions.every((o) => o.value === 0)}
+          className={styles.timezoneSelect}
         >
           {timeZoneOptions.map((o) => (
             <option key={o.key} value={o.value}>